rationalize Target::name, use Target::name for xDS
Makes Target::name the canonical name for all xDS, and start including
port everywhere. This makes all of the xDS semantics a little more
rigid, but is necessary if we want to be able to infer the difference
between a k8s Service target and a DNS target.

Thrown in is also some extra validation for DNS-style labels, which we
use for k8s Service names, and for DNS hostnames.