use a built wheel in smoke tests
use the maturin action to build a wheel and test that in our smoke
tests, instead of copying the wheel into our containers manually. this
gives us a slightly more realistic test of what folks will install/run
and means we can maybe re-use this in python CI as well.