junction-core: wait for cached results
wait for cached results when resolving routes/endpoints so that we don't
have to retry. adds a basic timeout mechanism.

to make this feasible, we completely changed the way that default routes
and backends are handled. instead of sometimes merging them, you can
now use defaults to completely replace dynamic configuration for
everything but endpoints. this removes any non-determinism and the
opportunity for confusion when figuring out what routes you've got -
you're either using a fully static set of routes or you're not.